requirements.txt
atlassian-python-api==3.22.0
PyGithub==1.55